@@ -1,14 +1,17 @@
 using System;
+using System.Text.Encodings.Web;
+using System.Text.Unicode;
 using Microsoft.AspNetCore.Builder;
 using Microsoft.AspNetCore.Hosting;
 using Microsoft.AspNetCore.Http;
 using Microsoft.AspNetCore.Localization;
 using Microsoft.AspNetCore.Mvc;
-using Microsoft.AspNetCore.Mvc.ApplicationModels;
 using Microsoft.AspNetCore.Rewrite;
 using Microsoft.AspNetCore.StaticFiles;
 using Microsoft.Extensions.Configuration;
 using Microsoft.Extensions.DependencyInjection;
+using Microsoft.Extensions.Hosting;
+using Microsoft.Extensions.WebEncoders;
 using Microsoft.WindowsAzure.Storage;
 
 namespace StyleEl
@@ -18,47 +21,52 @@ namespace StyleEl
 		public IConfiguration Configuration { get; }
 
 		public Startup(IConfiguration configuration)
-		{
-			Configuration = configuration;
-		}
+			=> Configuration = configuration;
 
 		public void ConfigureServices(IServiceCollection services)
 		{
-			services.AddMvc(ConfigureMvc)
-				.SetCompatibilityVersion(CompatibilityVersion.Latest);
+			services.AddControllers(ConfigureMvc);
+			services.AddRazorPages();
 			services.AddRouting(opt => opt.LowercaseUrls = true);
 			services.AddResponseCaching();
-			services.AddHttpsRedirection(opt => {
-				opt.HttpsPort = 443;
-				opt.RedirectStatusCode = StatusCodes.Status301MovedPermanently;
-			});
-			services.AddSingleton<IPageApplicationModelProvider, DefaultResponseCacheApplicationModelProvider>();
 			services.AddSingleton(CloudStorageAccount.Parse(Configuration.GetConnectionString("Storage")));
 			services.AddSingleton<MarkdownProvider>();
+			services.Configure<WebEncoderOptions>(opt =>
+			{
+				opt.TextEncoderSettings = new TextEncoderSettings(UnicodeRanges.All);
+			});
 		}
 
-		public void Configure(IApplicationBuilder app, IHostingEnvironment env)
+		public void Configure(IApplicationBuilder app, IWebHostEnvironment env)
 		{
 			if (!env.IsProduction())
-			{
 				app.UseDeveloperExceptionPage();
-			}
 			else
-			{
 				app.UseExceptionHandler("/Error");
-				app.UseHttpsRedirection();
-			}
 
-			app.UseRewriter(new RewriteOptions()
-				.AddRedirect("^index$", "/")
-				.AddRedirect("^(.+)/$", "$1"));
 			app.UseRequestLocalization(new RequestLocalizationOptions
 			{
 				DefaultRequestCulture = new RequestCulture("ru-RU")
 			});
 			app.UseStaticFiles(new StaticFileOptions { OnPrepareResponse = PrepareStaticFile });
+			app.UseRewriter(ConfigureRewrite(env));
 			app.UseResponseCaching();
-			app.UseMvc();
+			app.UseRouting();
+			app.UseEndpoints(endpoints =>
+			{
+				endpoints.MapControllers();
+				endpoints.MapRazorPages();
+			});
+		}
+
+		RewriteOptions ConfigureRewrite(IWebHostEnvironment env)
+		{
+			var opt = new RewriteOptions();
+			if (env.IsProduction())
+				opt.AddRedirectToHttpsPermanent();
+			return opt
+				.AddRedirect("^index$", "/")
+				.AddRedirect("^(.+)/$", "$1");
 		}
 
 		void ConfigureMvc(MvcOptions options)
