@@ -4,7 +4,7 @@ namespace StyleEl.Models
 {
 	public class Partner : TableEntity
 	{
-		public string Name { get; set; }
-		public string Url { get; set; }
+		public string Name { get; set; } = "";
+		public string? Url { get; set; }
 	}
 }
