@@ -6,8 +6,8 @@ namespace StyleEl.Models
 	public class Publication : TableEntity
 	{
 		public DateTime Date { get; set; }
-		public string Name { get; set; }
-		public string Place { get; set; }
-		public string Text { get; set; }
+		public string Name { get; set; } = "";
+		public string Place { get; set; } = "";
+		public string Text { get; set; } = "";
 	}
 }
