@@ -1,6 +1,6 @@
 ﻿@page
 @{
-	ViewBag.Title = "Контакты - «Стиль Л»";
+	ViewBag.Title = "Контакты - «Стиль Л»";
 	ViewBag.Header = "Контакты";
 }
 <h3 class="notop">Директор</h3>
