@@ -2,7 +2,7 @@
 @model IndexModel
 @inject MarkdownProvider Markdown
 @{
-	ViewBag.Title = "«Стиль Л» - дизайн, строительство, отделка";
+	ViewBag.Title = "«Стиль Л» - дизайн, строительство, отделка";
 }
 
 <div class="doc gallery">
